+++ src/de/jost_net/JVerein/gui/dialogs/TablePartExportDialog.java
@@ -1,5 +1,4 @@
 /**********************************************************************
- * Copyright (c) by Heiner Jostkleigrewe
  * This program is free software: you can redistribute it and/or modify it under the terms of the 
  * GNU General Public License as published by the Free Software Foundation, either version 3 of the 
  * License, or (at your option) any later version.
@@ -11,8 +10,6 @@
  * You should have received a copy of the GNU General Public License along with this program.  If not, 
  * see <http://www.gnu.org/licenses/>.
  * 
- * heiner@jverein.de
- * www.jverein.de
  **********************************************************************/
 package de.jost_net.JVerein.gui.dialogs;
 
@@ -32,7 +29,6 @@ import org.eclipse.swt.graphics.Color;
 import org.eclipse.swt.graphics.FontData;
 import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.widgets.Composite;
-import org.eclipse.swt.widgets.FileDialog;
 import org.eclipse.swt.widgets.TabFolder;
 import org.eclipse.swt.widgets.Table;
 import org.eclipse.swt.widgets.TableColumn;
@@ -48,19 +44,9 @@ import com.itextpdf.text.DocumentException;
 import com.itextpdf.text.Element;
 import com.itextpdf.text.Font;
 
-import de.jost_net.JVerein.Einstellungen;
-import de.jost_net.JVerein.Einstellungen.Property;
-import de.jost_net.JVerein.gui.input.FormularInput;
 import de.jost_net.JVerein.gui.parts.JVereinTablePart;
-import de.jost_net.JVerein.io.FileViewer;
 import de.jost_net.JVerein.io.Reporter;
-import de.jost_net.JVerein.keys.FormularArt;
 import de.jost_net.JVerein.rmi.Formular;
-import de.willuhn.jameica.gui.GUI;
-import de.willuhn.jameica.gui.dialogs.AbstractDialog;
-import de.willuhn.jameica.gui.input.CheckboxInput;
-import de.willuhn.jameica.gui.input.IntegerInput;
-import de.willuhn.jameica.gui.input.SelectInput;
 import de.willuhn.jameica.gui.parts.Button;
 import de.willuhn.jameica.gui.parts.ButtonArea;
 import de.willuhn.jameica.gui.parts.Column;
@@ -70,55 +56,15 @@ import de.willuhn.jameica.system.Settings;
 import de.willuhn.logging.Logger;
 import de.willuhn.util.ApplicationException;
 
-public class TablePartExportDialog extends AbstractDialog<Boolean>
+public class TablePartExportDialog extends AbstractPartExportDialog
 {
-  public enum ExportArt
-  {
-    PDF,
-    CSV
-  }
-
-  private boolean success = false;
-
-  private Settings settings;
-
   private Table table;
-
-  private String title;
-
-  private String subtitle;
-
-  private String filename;
-
-  private ExportArt art;
-
-  private String settingPrefix;
-
-  IntegerInput links;
-
-  IntegerInput rechts;
-
-  IntegerInput oben;
-
-  IntegerInput unten;
-
-  private CheckboxInput querformat;
-
-  private SelectInput vordergrund;
-
-  private SelectInput hintergrund;
-
-  private JVereinTablePart spaltenList;
-
-  private CheckboxInput headerTransparent;
-
-  private CheckboxInput zellenTransparent;
 
   public TablePartExportDialog(Table table, String settingPrefix, ExportArt art,
       String title, String subtitle, String filename)
       throws ApplicationException
   {
-    super(TablePartExportDialog.POSITION_CENTER);
+    super(settingPrefix, art, title, subtitle, filename);
 
     if (table == null || table.isDisposed() || !(table instanceof Table))
     {
@@ -130,15 +76,6 @@ public class TablePartExportDialog extends AbstractDialog<Boolean>
     }
 
     this.table = table;
-    this.title = title;
-    this.subtitle = subtitle;
-    this.filename = filename;
-    this.art = art;
-    this.settingPrefix = settingPrefix + art.toString() + ".";
-
-    setTitle("Tabelle exportieren");
-    setSize(400, SWT.DEFAULT);
-
     settings = new Settings(this.getClass());
   }
 
@@ -225,49 +162,7 @@ public class TablePartExportDialog extends AbstractDialog<Boolean>
       }, null, false, "eraser.png"));
       tabSpalten.addButtonArea(buttons);
 
-      links = new IntegerInput(settings.getInt(settingPrefix + "links", 20));
-      rechts = new IntegerInput(settings.getInt(settingPrefix + "rechts", 20));
-      oben = new IntegerInput(settings.getInt(settingPrefix + "oben", 20));
-      unten = new IntegerInput(settings.getInt(settingPrefix + "unten", 20));
-      tabRaender.addLabelPair("Links", links);
-      tabRaender.addLabelPair("Rechts", rechts);
-      tabRaender.addLabelPair("Oben", oben);
-      tabRaender.addLabelPair("Unten", unten);
-
-      // IntegerInput links2 = new IntegerInput(settings.getInt(id + "links2",
-      // 20));
-      // IntegerInput rechts2 = new IntegerInput(
-      // settings.getInt(id + "rechts2", 20));
-      // IntegerInput oben2 = new IntegerInput(settings.getInt(id + "oben2",
-      // 20));
-      // IntegerInput unten2 = new IntegerInput(settings.getInt(id + "unten2",
-      // 20));
-      // tabRaender.addLabelPair("Links ab 2. Seite", links2);
-      // tabRaender.addLabelPair("Rechts ab 2. Seite", rechts2);
-      // tabRaender.addLabelPair("Oben ab 2. Seite", oben2);
-      // tabRaender.addLabelPair("Unten ab 2. Seite", unten2);
-
-      hintergrund = new FormularInput(FormularArt.HINTERGRUND,
-          settings.getString(settingPrefix + "hintergrund", ""));
-      hintergrund.setPleaseChoose("Kein Formular");
-      vordergrund = new FormularInput(FormularArt.HINTERGRUND,
-          settings.getString(settingPrefix + "vordergrund", ""));
-      vordergrund.setPleaseChoose("Kein Formular");
-      headerTransparent = new CheckboxInput(settings.getBoolean(
-          settingPrefix + "headerTransparent", (Boolean) Einstellungen
-              .getEinstellung(Property.TABELLEN_HEADER_TRANSPARENT)));
-      zellenTransparent = new CheckboxInput(settings.getBoolean(
-          settingPrefix + "zellenTransparent", (Boolean) Einstellungen
-              .getEinstellung(Property.TABELLEN_ZELLEN_TRANSPARENT)));
-      querformat = new CheckboxInput(
-          settings.getBoolean(settingPrefix + "quer", false));
-      tabFormular.addLabelPair("Formular Hintergrund", hintergrund);
-      tabFormular.addLabelPair("Formular Vordergrund", vordergrund);
-      tabFormular.addLabelPair("Tabellen Header transparent",
-          headerTransparent);
-      tabFormular.addLabelPair("Tabellen Zellen transparent",
-          zellenTransparent);
-      tabFormular.addLabelPair("Querformat", querformat);
+      addRaenderFormularTabs(tabRaender, tabFormular);
     }
     else
     {
@@ -290,69 +185,8 @@ public class TablePartExportDialog extends AbstractDialog<Boolean>
     b.paint(parent);
   }
 
-  private void export() throws ApplicationException
-  {
-    try
-    {
-      String extension = "";
-      switch (art)
-      {
-        case CSV:
-          extension = ".csv";
-          break;
-        case PDF:
-          extension = ".pdf";
-          break;
-      }
-
-      FileDialog fd = new FileDialog(GUI.getShell(), SWT.SAVE);
-      fd.setText("Ausgabedatei wählen.");
-
-      String path = settings.getString(settingPrefix + "lastdir",
-          System.getProperty("user.home"));
-      if (path != null && path.length() > 0)
-      {
-        fd.setFilterPath(path);
-      }
-
-      fd.setFileName(filename);
-      fd.setFilterExtensions(new String[] { "*" + extension });
-
-      final String p = fd.open();
-
-      if (p == null || p.length() == 0)
-      {
-        throw new OperationCanceledException("Abgebrochen");
-      }
-
-      File file = new File(p);
-      settings.setAttribute(settingPrefix + "lastdir", file.getParent());
-
-      switch (art)
-      {
-        case CSV:
-          exportCSV(file);
-          break;
-        case PDF:
-          exportPDF(file);
-          break;
-      }
-      saveSettings();
-
-      FileViewer.show(file);
-
-      success = true;
-      close();
-    }
-    catch (IOException | DocumentException e)
-    {
-      String fehler = "Fehler beim Export";
-      Logger.error(fehler, e);
-      throw new ApplicationException(fehler);
-    }
-  }
-
-  private void exportCSV(File file) throws IOException
+  @Override
+  protected void exportCSV(File file) throws IOException
   {
     try (ICsvMapWriter writer = new CsvMapWriter(new FileWriter(file),
         CsvPreference.EXCEL_NORTH_EUROPE_PREFERENCE))
@@ -387,7 +221,8 @@ public class TablePartExportDialog extends AbstractDialog<Boolean>
     }
   }
 
-  private void exportPDF(File file) throws IOException, DocumentException
+  @Override
+  protected void exportPDF(File file) throws IOException, DocumentException
   {
     try (FileOutputStream fos = new FileOutputStream(file);
         Reporter reporter = new Reporter(fos, title, subtitle,
@@ -456,7 +291,7 @@ public class TablePartExportDialog extends AbstractDialog<Boolean>
   }
 
   @SuppressWarnings("unchecked")
-  private void saveSettings() throws RemoteException
+  protected void saveSettings() throws RemoteException
   {
     List<TableColumn> itemsChecked = spaltenList.getItems();
     for (TableColumn col : (List<TableColumn>) spaltenList.getItems(false))
@@ -469,38 +304,7 @@ public class TablePartExportDialog extends AbstractDialog<Boolean>
             (Integer) col.getData());
       }
     }
-
-    if (art.equals(ExportArt.PDF))
-    {
-      settings.setAttribute(settingPrefix + "links",
-          (Integer) links.getValue());
-      settings.setAttribute(settingPrefix + "rechts",
-          (Integer) rechts.getValue());
-      settings.setAttribute(settingPrefix + "oben", (Integer) oben.getValue());
-      settings.setAttribute(settingPrefix + "unten",
-          (Integer) unten.getValue());
-
-      settings.setAttribute(settingPrefix + "hintergrund",
-          hintergrund.getValue() == null ? null
-              : ((Formular) hintergrund.getValue()).getID());
-      settings.setAttribute(settingPrefix + "vordergrund",
-          vordergrund.getValue() == null ? null
-              : ((Formular) vordergrund.getValue()).getID());
-
-      settings.setAttribute(settingPrefix + "headerTransparent",
-          (Boolean) headerTransparent.getValue());
-      settings.setAttribute(settingPrefix + "zellenTransparent",
-          (Boolean) zellenTransparent.getValue());
-
-      settings.setAttribute(settingPrefix + "quer",
-          (Boolean) querformat.getValue());
-    }
-  }
-
-  @Override
-  protected Boolean getData() throws Exception
-  {
-    return success;
+    super.saveSettings();
   }
 
 }
