+++ src/de/jost_net/JVerein/gui/view/ArbeitseinsatzListeView.java
@@ -22,7 +22,7 @@ import de.jost_net.JVerein.gui.action.ArbeitseinsatzZusatzbetraegeAction;
 import de.jost_net.JVerein.gui.action.DokumentationAction;
 import de.jost_net.JVerein.gui.action.NewAction;
 import de.jost_net.JVerein.gui.control.ArbeitseinsatzControl;
-import de.jost_net.JVerein.gui.dialogs.TablePartExportDialog.ExportArt;
+import de.jost_net.JVerein.gui.dialogs.AbstractPartExportDialog.ExportArt;
 import de.jost_net.JVerein.gui.parts.ToolTipButton;
 import de.jost_net.JVerein.rmi.Arbeitseinsatz;
 import de.willuhn.jameica.gui.AbstractView;
